git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/branches/hpx/release@7292 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ external/atomic/boost/atomic/detail/builder.hpp
@@ -344,14 +344,14 @@ private:
         address&=~(sizeof(larger_integral_type)-1);
         return *reinterpret_cast<Base *>(address);
     }
-    unsigned int get_offset(void) const volatile
+    intptr_t get_offset(void) const volatile
     {
         intptr_t address=(intptr_t)this;
         address&=(sizeof(larger_integral_type)-1);
         return address;
     }
 
-    unsigned int get_shift(void) const volatile
+    intptr_t get_shift(void) const volatile
     {
 #if defined(BOOST_LITTLE_ENDIAN)
         return get_offset()*8;
